+++ src/app/shared/shared.module.ts
@@ -3,19 +3,22 @@ import { CommonModule } from '@angular/common';
 import { HighlightDirective } from './directives/highlight.directive';
 import { WheelDirective } from './directives/wheel.directive';
 import { ChangeSizeDirective } from './directives/change-size.directive';
+import { OrderByPipe } from './pipes/order-by.pipe';
 
 @NgModule({
   declarations: [
     HighlightDirective,
     WheelDirective,
-    ChangeSizeDirective
+    ChangeSizeDirective,
+    OrderByPipe
   ],
   imports: [
     CommonModule
   ],
   exports: [
     HighlightDirective,
-    WheelDirective
+    WheelDirective,
+    OrderByPipe
   ]
 })
 export class SharedModule { }